Separate dangerous parts into classwrapperauto
Separate dangerous parts into classwrapperauto

minor
